fix how the config is loaded from a path and merge the ops/ctl config tests
