+++ docker-compose.yml
@@ -17,7 +17,7 @@ services:
     expose:
       - 8080
     environment:
-      FIDES__CONFIG_PATH: "/fides/.fides/"
+      FIDES__CONFIG_PATH: "/fides/.fides/fides.toml"
       FIDES__CLI__ANALYTICS_ID: ${FIDES__CLI__ANALYTICS_ID}
       FIDES__CLI__SERVER_HOST: "fides"
       FIDES__CLI__SERVER_PORT: "8080"
